services/notification_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy>=2.0
mysqlclient>=2.2
python-dotenv>=1.0
httpx>=0.28
python-jose[cryptography]>=3.3
celery>=5.4.0
redis>=5.2.0
